+++ firmware/prototyper/src/riscv/mod.rs
@@ -1,3 +1,4 @@
+pub mod allwinner_v861;
 pub mod csr;
 pub mod spacemit_k1;